requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
diskcache==5.6.3
ExifRead==3.0.0
fastapi==0.103.2
ffmpeg-python==0.2.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
macholib==1.16.3
MarkupSafe==2.1.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pydantic==2.4.2
pydantic_core==2.10.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzlocal==5.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==3.0.0
